chore: make hannwindow accept size_t

+++ packages/react-native-executorch/common/rnexecutorch/data_processing/dsp.cpp
@@ -9,10 +9,10 @@ namespace rnexecutorch::dsp {
 
 using namespace rnexecutorch::dsp;
 
-std::vector<float> hannWindow(int size) {
+std::vector<float> hannWindow(size_t size) {
   // https://www.mathworks.com/help/signal/ref/hann.html
   std::vector<float> window(size);
-  for (int i = 0; i < size; i++) {
+  for (size_t i = 0; i < size; i++) {
     window[i] = 0.5f * (1 - std::cos(2 * M_PI * i / size));
   }
   return window;
